Add and reorganize tests for master shops and assets
Introduced a dedicated test suite for Master Shops in `MastersTest.php`, including creation, update, and UI validation scenarios. Removed related test cases from `GoodsTest.php` to maintain separation of concerns. Updated role counts in SysAdmin tests to reflect new logic.
